fix: add customer creation logic in Channel Partner doctype
